+++ src/main/java/io/github/thebusybiscuit/slimefun4/api/items/groups/LockedItemGroup.java
@@ -87,7 +87,7 @@ public class LockedItemGroup extends ItemGroup {
             }
         }
 
-        for (ItemGroup category : Slimefun.getRegistry().getCategories()) {
+        for (ItemGroup category : Slimefun.getRegistry().getAllItemGroups()) {
             if (namespacedKeys.remove(category.getKey())) {
                 addParent(category);
             }