+++ src/main/java/io/github/thebusybiscuit/slimefun4/api/items/groups/NestedItemGroup.java
@@ -65,7 +65,7 @@ public class NestedItemGroup extends FlexItemGroup {
     @Override
     @ParametersAreNonnullByDefault
     public boolean isVisible(Player p, PlayerProfile profile, SlimefunGuideMode mode) {
-        return true;
+        return mode == SlimefunGuideMode.SURVIVAL_MODE;
     }
 
     @Override