+++ src/main/java/io/github/thebusybiscuit/slimefun4/core/guide/GuideHistory.java
@@ -189,7 +189,7 @@ public class GuideHistory {
         if (entry == null) {
             guide.openMainMenu(profile, mainMenuPage);
         } else if (entry.getIndexedObject() instanceof ItemGroup) {
-            guide.openCategory(profile, (ItemGroup) entry.getIndexedObject(), entry.getPage());
+            guide.openItemGroup(profile, (ItemGroup) entry.getIndexedObject(), entry.getPage());
         } else if (entry.getIndexedObject() instanceof SlimefunItem) {
             guide.displayItem(profile, (SlimefunItem) entry.getIndexedObject(), false);
         } else if (entry.getIndexedObject() instanceof ItemStack) {